requirements.txt
# E-commerce Personalization Platform - Requirements

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2
pydantic-settings==2.0.3

# OpenAI Integration
openai==1.3.5

# Vector Databases
faiss-cpu==1.7.4
# For Pinecone: uncomment below and comment out faiss
# pinecone-client==3.0.2

# Search and Database
elasticsearch==8.10.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Payment Integration
stripe==7.2.0

# AWS Integration
boto3==1.28.84
botocore==1.31.84

# HTTP Client
httpx==0.25.1

# Data Processing
numpy==1.24.3
pandas==2.1.3

# Caching
redis==5.0.1

# Environment and Config
python-dotenv==1.0.0

# Logging and Monitoring
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx-mock==0.25.0

# Development
black==23.12.0
flake8==6.1.0
mypy==1.7.1
isort==5.13.2

# Production Server
gunicorn==21.2.0
